@@ -1,12 +1,29 @@
 module github.com/jasonstubblefield/ingredients
 
-go 1.12
+go 1.21
+
+toolchain go1.22.2
 
 require (
+	github.com/astappiev/microdata v1.0.2
 	github.com/goccy/go-json v0.9.11
 	github.com/jinzhu/inflection v1.0.0
 	github.com/schollz/logger v1.2.0
 	github.com/schollz/progressbar/v3 v3.7.6
-	github.com/stretchr/testify v1.3.0
-	golang.org/x/net v0.0.0-20200707034311-ab3426394381
+	github.com/stretchr/testify v1.9.0
+	golang.org/x/net v0.26.0
+)
+
+require (
+	github.com/astappiev/fixjson v1.0.1 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mattn/go-runewidth v0.0.10 // indirect
+	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	golang.org/x/crypto v0.24.0 // indirect
+	golang.org/x/sys v0.21.0 // indirect
+	golang.org/x/term v0.21.0 // indirect
+	golang.org/x/text v0.16.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
